+++ src/vertex_buffer.rs
@@ -184,7 +184,7 @@ impl<T: Send + Copy> VertexBuffer<T> {
     ///
     #[experimental]
     pub unsafe fn new_raw(display: &super::Display, data: Vec<T>,
-                          bindings: VertexFormat, elements_size: uint) -> VertexBuffer<T>
+                          bindings: VertexFormat, elements_size: usize) -> VertexBuffer<T>
     {
         VertexBuffer {
             buffer: VertexBufferAny {
@@ -253,7 +253,7 @@ impl<T: Send + Copy> VertexBuffer<T> {
     ///
     /// Only available if the `gl_extensions` feature is enabled.
     #[cfg(feature = "gl_extensions")]
-    pub fn read_slice(&self, offset: uint, size: uint) -> Vec<T> {
+    pub fn read_slice(&self, offset: usize, size: usize) -> Vec<T> {
         self.buffer.buffer.read_slice::<buffer::ArrayBuffer, T>(offset, size)
     }
 
@@ -268,7 +268,7 @@ impl<T: Send + Copy> VertexBuffer<T> {
 
 impl<T> VertexBuffer<T> {
     /// Returns the number of bytes between two consecutive elements in the buffer.
-    pub fn get_elements_size(&self) -> uint {
+    pub fn get_elements_size(&self) -> usize {
         self.buffer.elements_size
     }
 
@@ -306,12 +306,12 @@ impl<'a, T> IntoVerticesSource<'a> for &'a VertexBuffer<T> {
 pub struct VertexBufferAny {
     buffer: Buffer,
     bindings: VertexFormat,
-    elements_size: uint,
+    elements_size: usize,
 }
 
 impl VertexBufferAny {
     /// Returns the number of bytes between two consecutive elements in the buffer.
-    pub fn get_elements_size(&self) -> uint {
+    pub fn get_elements_size(&self) -> usize {
         self.elements_size
     }
 
@@ -409,7 +409,7 @@ pub enum AttributeType {
 ///
 /// The first element is the name of the binding, the second element is the offset
 /// from the start of each vertex to this element, and the third element is the type.
-pub type VertexFormat = Vec<(String, uint, AttributeType)>;
+pub type VertexFormat = Vec<(String, usize, AttributeType)>;
 
 /// Trait for structures that represent a vertex.
 ///
